requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12
djangorestframework-simplejwt>=5.2.2
Pillow>=9.1.0
django-cors-headers>=3.13